wpiutil: Rename Optional.h to LLVMOptional.h
